Expression: Move expression-parser tests to separate sub-package
